Code/requirements.txt
dnspython==2.3.0
gdown==4.7.1
matplotlib==3.7.1
netifaces==0.11.0
numpy==1.24.3
pandas==2.0.2
scikit_learn==1.2.2
Twisted==22.10.0
